Feat/additional issue templates (#7)
* chore: updates bug report template

Added priority level section and report name field to the bug report template. Removed the default assignee to make assignments more flexible.

Refs: [#5](https://github.com/fernandonevarez/lms/issues/5)

* chore: add documentation update issue template

Created a new issue template to streamline and standardize the process for requesting documentation updates.

Refs: [#5](https://github.com/fernandonevarez/lms/issues/5)

* chore: updates new feature issue template

Updated the format and default labels of the new feature issue template to improve clarity and categorization.

Refs: [#5](https://github.com/fernandonevarez/lms/issues/5)

* chore: add hotfix issue template

Created a new issue template for hotfixes to ensure quick and consistent reporting of urgent issues.

Refs: [#5](https://github.com/fernandonevarez/lms/issues/5)

* chore: add refactor request issue template

Created a new issue template for refactor requests to standardize how code improvement suggestions are made.

Refs: [#5](https://github.com/fernandonevarez/lms/issues/5)

* fix: correct type for Priority Level in issue templates

Fixed a bug in the issue templates where the type was incorrectly set to 'select'. Changed it to 'dropdown' as 'select' is not a supported type.

---------

Co-authored-by: David Nevarez <69279785+SloppyFlipFlop@users.noreply.github.com>