Dev (#10)
* style: update project font for improved design

This commit updates the project's font to enhance the visual design and readability across the application.

* feat(auth): create layout file for authentication paths

This commit creates a dedicated layout file for authentication-related paths, providing a unique UI design tailored for user authentication related pages.

* feat(layout): create layout file for main project pages

This commit creates a layout file for the main pages of the project, ensuring a uniform UI design across all pages within the main route.

* feat(seo): add basic page metadata for SEO and page content

Added page's metadata for improved SEO. Also updated initial page content.

* chore: add @types/node-cron to devDependencies

Imported @types/node-cron as a development dependency to support TypeScript type definitions for node-cron.

* feat(styles): add global project colors, font family, and custom transition duration to Tailwind config

Configured Tailwind CSS to include global project colors, font family settings, and a custom transition duration for consistent styling across the project.

* style: update project's global font family

Updated the global font family in Tailwind configuration for a more consistent and modern look across the project.

* feat: update font family and add metadata for SEO

Updated the global font family for improved consistency. Added meta tags including title, description, and keywords to enhance SEO.

* chore: update project's favicon

Replaced the existing favicon with a new one for better brand representation.

* feat: add manifest file for project

Created a manifest file to provide metadata for the project, including app name and theme colors for better integration with devices and browsers.

* feat: add robots file

Added robots.ts file to manage and configure web crawling rules programmatically.

* feat: add sitemap.ts for generating sitemap

Added sitemap.ts file to programmatically generate a sitemap for the project, improving SEO and site structure for search engines.

* feat(seo): add metadata for improved individual page SEO

Enhanced individual pages with metadata, including titles, descriptions, and keywords, to improve SEO and search engine visibility.

* feat(seo): add metadata to auth route layout for improved SEO

Added metadata to the layout file of the auth route to enhance SEO for the authentication pages.

* feat: create basic footer component

Implemented a basic footer component for the project as a placeholder for now until i create an actual one for the project.

* feat(api): add route to regenerate sitemap

Created an API route to programmatically regenerate the sitemap for the project, ensuring up-to-date site structure for SEO purposes.

* refactor: improve code structure and readability of file

Refactored the file to enhance code structure, readability, and maintainability.

* Feat/additional issue templates (#7)

* chore: updates bug report template

Added priority level section and report name field to the bug report template. Removed the default assignee to make assignments more flexible.

Refs: [#5](https://github.com/fernandonevarez/lms/issues/5)

* chore: add documentation update issue template

Created a new issue template to streamline and standardize the process for requesting documentation updates.

Refs: [#5](https://github.com/fernandonevarez/lms/issues/5)

* chore: updates new feature issue template

Updated the format and default labels of the new feature issue template to improve clarity and categorization.

Refs: [#5](https://github.com/fernandonevarez/lms/issues/5)

* chore: add hotfix issue template

Created a new issue template for hotfixes to ensure quick and consistent reporting of urgent issues.

Refs: [#5](https://github.com/fernandonevarez/lms/issues/5)

* chore: add refactor request issue template

Created a new issue template for refactor requests to standardize how code improvement suggestions are made.

Refs: [#5](https://github.com/fernandonevarez/lms/issues/5)

* fix: correct type for Priority Level in issue templates

Fixed a bug in the issue templates where the type was incorrectly set to 'select'. Changed it to 'dropdown' as 'select' is not a supported type.

---------

Co-authored-by: David Nevarez <69279785+SloppyFlipFlop@users.noreply.github.com>

* chore: update domain from '.org' to '.io' (#9)

Updated all instances of the '.org' domain to '.io'

refs: [#8](https://github.com/fernandonevarez/lms/issues/8).

---------

Co-authored-by: David Nevarez <69279785+SloppyFlipFlop@users.noreply.github.com>